requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.9.85
botocore==1.12.85
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.2
django-crispy-forms==1.7.2
django-filter==2.4.0a
django-storages==1.7.1
djangorestframework==3.12.1
docutils==0.14
gunicorn==19.9.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.2.0
psycopg2-binary==2.8.6
python-dateutil==2.7.5
pytz==2018.9
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.1.13
simplejson==3.17.2
six==1.12.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.24.1
wincertstore==0.2
